[Rename] QueryParam, params 폴더이동
역할을 명확하게 부여하기 위해 model에서 util로의 이동
