+++ src/main/java/util/QueryParams.java
@@ -1,4 +1,4 @@
-package model;
+package util;
 
 import java.util.Arrays;
 import java.util.Map;
